rename HOCs
